third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-compileScript-expected.txt

{
    exceptionDetails : {
        columnNumber : 2
        lineNumber : 1
        scriptId : 0
        text : Uncaught SyntaxError: Unexpected end of input
        url : foo1.js
    }
}
{
    scriptId : 0
}
{
}
{
    exceptionDetails : {
        columnNumber : 13
        lineNumber : 0
        scriptId : 0
        text : Uncaught SyntaxError: Unexpected identifier
        url : foo4.js
    }
}
